refactor: main components - 로딩 및 에러 처리 필요
