+++ src/components/Main/Products/Title.tsx
@@ -1,0 +1,8 @@
+/** @jsxImportSource @emotion/react */
+import { title } from './BestProducts.styles';
+
+const Title: React.FC<{ children: React.ReactNode }> = ({ children }) => {
+  return <h2 css={title}>{children}</h2>;
+};
+
+export default Title;
